Plot function
A general function for plotting, such that all the plots in the report will look consitent. The function takes a two dimensional array for both x and y in case you want to plot multiple lines. Furthermore, you need to input a list of strings for the labels and give the function a filename